requirements.txt
streamlit
llama-parse
llama-index
llama-index-core
llama-index-llms-azure-openai
llama-index-embeddings-openai
python-dotenv
